move validate localityLbsetting to validation.go
