+++ pilot/pkg/kube/inject/webhook.go
@@ -36,6 +36,7 @@ import (
 
 	meshconfig "istio.io/api/mesh/v1alpha1"
 	"istio.io/istio/pilot/cmd"
+	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pkg/log"
 )
 
@@ -87,7 +88,7 @@ func loadConfig(injectFile, meshFile string) (*Config, *meshconfig.MeshConfig, e
 	if err != nil {
 		return nil, nil, err
 	}
-	err = cmd.ValidateMeshConfig(meshConfig)
+	err = model.ValidateMeshConfig(meshConfig)
 	if err != nil {
 		return nil, nil, err
 	}
