+++ pilot/pkg/model/validation_test.go
@@ -3996,3 +3996,171 @@ func TestValidateSidecar(t *testing.T) {
 		})
 	}
 }
+
+func TestValidateLocalityLbSetting(t *testing.T) {
+	cases := []struct {
+		name  string
+		in    *meshconfig.LocalityLoadBalancerSetting
+		valid bool
+	}{
+		{
+			name:  "valid mesh config without LocalityLoadBalancerSetting",
+			in:    nil,
+			valid: true,
+		},
+
+		{
+			name: "invalid LocalityLoadBalancerSetting_Distribute total weight > 100",
+			in: &meshconfig.LocalityLoadBalancerSetting{
+				Distribute: []*meshconfig.LocalityLoadBalancerSetting_Distribute{
+					{
+						From: "a/b/c",
+						To: map[string]uint32{
+							"a/b/c": 80,
+							"a/b1":  25,
+						},
+					},
+				},
+			},
+			valid: false,
+		},
+		{
+			name: "invalid LocalityLoadBalancerSetting_Distribute total weight < 100",
+			in: &meshconfig.LocalityLoadBalancerSetting{
+				Distribute: []*meshconfig.LocalityLoadBalancerSetting_Distribute{
+					{
+						From: "a/b/c",
+						To: map[string]uint32{
+							"a/b/c": 80,
+							"a/b1":  15,
+						},
+					},
+				},
+			},
+			valid: false,
+		},
+		{
+			name: "invalid LocalityLoadBalancerSetting_Distribute weight = 0",
+			in: &meshconfig.LocalityLoadBalancerSetting{
+				Distribute: []*meshconfig.LocalityLoadBalancerSetting_Distribute{
+					{
+						From: "a/b/c",
+						To: map[string]uint32{
+							"a/b/c": 0,
+							"a/b1":  100,
+						},
+					},
+				},
+			},
+			valid: false,
+		},
+		{
+			name: "invalid LocalityLoadBalancerSetting specify both distribute and failover",
+			in: &meshconfig.LocalityLoadBalancerSetting{
+				Distribute: []*meshconfig.LocalityLoadBalancerSetting_Distribute{
+					{
+						From: "a/b/c",
+						To: map[string]uint32{
+							"a/b/c": 80,
+							"a/b1":  20,
+						},
+					},
+				},
+				Failover: []*meshconfig.LocalityLoadBalancerSetting_Failover{
+					{
+						From: "region1",
+						To:   "region2",
+					},
+				},
+			},
+			valid: false,
+		},
+
+		{
+			name: "invalid failover src and dst have same region",
+			in: &meshconfig.LocalityLoadBalancerSetting{
+				Failover: []*meshconfig.LocalityLoadBalancerSetting_Failover{
+					{
+						From: "region1",
+						To:   "region1",
+					},
+				},
+			},
+			valid: false,
+		},
+	}
+
+	for _, c := range cases {
+		if got := validateLocalityLbSetting(c.in); (got == nil) != c.valid {
+			t.Errorf("ValidateLocalityLbSetting failed on %v: got valid=%v but wanted valid=%v: %v",
+				c.name, got == nil, c.valid, got)
+		}
+	}
+}
+
+func TestValidateLocalities(t *testing.T) {
+	cases := []struct {
+		name       string
+		localities []string
+		valid      bool
+	}{
+		{
+			name:       "multi wildcard locality",
+			localities: []string{"*/zone/*"},
+			valid:      false,
+		},
+		{
+			name:       "wildcard not in suffix",
+			localities: []string{"*/zone"},
+			valid:      false,
+		},
+		{
+			name:       "explicit wildcard region overlap",
+			localities: []string{"*", "a/b/c"},
+			valid:      false,
+		},
+		{
+			name:       "implicit wildcard region overlap",
+			localities: []string{"a", "a/b/c"},
+			valid:      false,
+		},
+		{
+			name:       "explicit wildcard zone overlap",
+			localities: []string{"a/*", "a/b/c"},
+			valid:      false,
+		},
+		{
+			name:       "implicit wildcard zone overlap",
+			localities: []string{"a/b", "a/b/c"},
+			valid:      false,
+		},
+		{
+			name:       "explicit wildcard subzone overlap",
+			localities: []string{"a/b/*", "a/b/c"},
+			valid:      false,
+		},
+		{
+			name:       "implicit wildcard subzone overlap",
+			localities: []string{"a/b", "a/b/c"},
+			valid:      false,
+		},
+		{
+			name:       "valid localities",
+			localities: []string{"a1/*", "a2/*", "a3/b3/c3", "a4/b4", "a5/b5/*"},
+			valid:      true,
+		},
+	}
+	for _, c := range cases {
+		t.Run(c.name, func(t *testing.T) {
+			err := validateLocalities(c.localities)
+			if !c.valid && err == nil {
+				t.Errorf("expect invalid localities")
+			}
+
+			if c.valid && err != nil {
+				t.Errorf("expect valid localities. but got err %v", err)
+			}
+		})
+	}
+
+}
